export display

+++ migrations/20170929001952_add_jobs.js
@@ -7,7 +7,7 @@ exports.up = function(knex, Promise) {
       table.string('note');
       table.integer('user_id').unsigned();
       table.dateTime('date');
-      table.integer('hour');
+      table.double('hour');
       table.timestamps();
       table.foreign('user_id').references('users.id')
     })
